Fix more conversion warnings

+++ include/viua/program.h
@@ -150,7 +150,7 @@ class Program {
      *  size of the program.
      */
     Program& calculateBranches(unsigned offset = 0); // FIXME: is unused, scheduled for removal
-    Program& calculateJumps(std::vector<std::tuple<int, int> >);
+    Program& calculateJumps(std::vector<std::tuple<uint64_t, uint64_t> >);
     std::vector<uint64_t> jumps();
     std::vector<uint64_t> jumpsAbsolute();
